refactor: make the wire-model module private (objects.py -> _objects.py)
_objects holds the raw pydantic wire models. It was never in any __all__ but was
still reachable as `smartschool.objects`, so `smartschool.objects.Result` (the
detail-less pydantic model) sat alongside the public session-aware `Result`.
Rename to _objects - matching the existing _xml_interface / _dev_tracing
convention - so only the session-aware classes (results.Result etc.) are public.
Updates all imports and regenerates the .pyi stubs.
