+++ src/smartschool/periods.py
@@ -1,7 +1,7 @@
 from collections.abc import Iterator
 from dataclasses import dataclass
 
-from .objects import Period
+from ._objects import Period
 from .session import SessionMixin
 
 __all__ = ["Periods"]
