+++ src/smartschool/reports.pyi
@@ -2,8 +2,8 @@
 from collections.abc import Iterable, Iterator
 from datetime import datetime
 
-from . import objects
-from .objects import SchoolClass
+from . import _objects as objects
+from ._objects import SchoolClass
 from .session import SessionMixin, Smartschool
 
 class Report(objects.Report, SessionMixin):
