+++ src/smartschool/student_support.py
@@ -1,7 +1,7 @@
 from collections.abc import Iterable, Iterator
 from dataclasses import dataclass
 
-from .objects import StudentSupportLink
+from ._objects import StudentSupportLink
 from .session import SessionMixin
 
 __all__ = ["StudentSupportLinks"]
