+++ tests/agenda_future_tasks_tests.py
@@ -4,7 +4,7 @@ from typing import TYPE_CHECKING
 from smartschool import FutureTasks, Smartschool
 
 if TYPE_CHECKING:
-    from smartschool.objects import FutureTaskOneDay
+    from smartschool._objects import FutureTaskOneDay
 
 
 def test_future_tasks_normal_flow(session: Smartschool):
